Hiding data constructors of `StandardVersion`
Data type `StandardVersion` was used to tag binary representation of
previous versions of the Dhall Standard.  We still need to keep this
data type around to support older versions. Starting with version 6.0.0
of the standard this tag is no longer required for encoded expressions.

Dhall API uses `StandardVersion` in various places.  Instead of removing
it entirely from the public API this commit opted to hide data
constructors and expose only `defaultStandardVersion`.

Dhall command line tool allowed `StandardVersion` to be specified on the
command line. After this commit it is no longer possible to do so.

issue dhall-lang/dhall-haskell#848